pyproject.toml
[project]
name = "vehicle_keypoints"
version = "0.1.0"
description = "Production-grade vehicle keypoint detection (14 anatomical car keypoints, CarFusion)."
requires-python = ">=3.13"
authors = [
  { name = "Daniil Kiselyov", email = "your-email@example.com" },
]
license = { text = "MIT" }
readme = "README.md"

dependencies = [
  "hydra-core>=1.3.2",
  "omegaconf>=2.3.0",
  "pyyaml>=6.0.1",
  "structlog>=24.1.0",
  "rich>=13.7.0",
  "jinja2>=3.1.0",
  "numpy>=1.26.0",
"torch>=2.2.0",
  "torchvision>=0.17.0",
  "lightning>=2.2.0",
  "torchmetrics>=1.3.0",
"ultralytics>=8.3.0",
  "pycocotools>=2.0.7",
]

[dependency-groups]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=4.1.0",
  "ruff>=0.15.10",
  "mypy>=1.20.1",
  "pre-commit>=4.0.0",
  "nbqa>=1.8.0",
  "nbstripout>=0.9.1",
  "ipykernel>=6.29.0",
  "matplotlib>=3.8.0",
  "deptry>=0.25.1",
  "bandit[toml]>=1.9.4",
  "interrogate>=1.7.0",
  "codespell>=2.4.2",
]
serving = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.29.0",
  "pydantic>=2.6.0",
  "python-multipart>=0.0.9",
  "prometheus-fastapi-instrumentator>=7.0.0",
  "huggingface-hub>=0.22.0",
"pillow>=10.2.0",
]
tracking = [
  "mlflow>=2.11.0",
  "dvc>=3.48.0",
  "wandb>=0.16.0",
]
docs = [
  "mkdocs-material>=9.5.0",
  "mkdocstrings[python]>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vehicle_keypoints"]

[tool.ruff]
line-length = 100
target-version = "py313"
extend-exclude = ["notebooks"]

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM", "RUF", "N"]
ignore = ["N812"]

[tool.ruff.lint.per-file-ignores]
"src/vehicle_keypoints/serving/routes.py" = ["B008"]

[tool.mypy]
python_version = "3.13"
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = "vehicle_keypoints.serving.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "vehicle_keypoints.data.datamodule"
ignore_errors = true
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short --cov=src --cov-report=term --cov-report=xml --cov-fail-under=10"

[tool.coverage.run]
source = ["src/vehicle_keypoints"]
branch = true

[tool.coverage.report]
exclude_lines = ["pragma: no cover", "if TYPE_CHECKING:", "raise NotImplementedError"]

[tool.bandit]
exclude_dirs = ["tests", "notebooks", "data", "artifacts"]
# B101: assert_used (pytest convention)
# B614: pytorch_load — we load only our own trained checkpoints
# B615: huggingface_unsafe_download — we download only our own / trusted HF repos
# B701: jinja2_autoescape_false — templates render Markdown/YAML, not HTML
skips = ["B101", "B614", "B615", "B701"]

[tool.interrogate]
ignore-init-method = true
ignore-init-module = true
ignore-magic = true
ignore-private = false
ignore-nested-functions = true
ignore-module = false
fail-under = 50
exclude = ["tests", "notebooks", "docs", "build", "setup.py"]
verbose = 1
color = true

[tool.deptry]
extend_exclude = ["notebooks", "tests"]
known_first_party = ["vehicle_keypoints"]
# DEP004: dependency-groups (serving, tracking, docs) are intentional optional installs
ignore = ["DEP004"]

[tool.deptry.per_rule_ignores]
DEP002 = ["nbstripout", "ipykernel", "pyyaml", "rich"]

[tool.codespell]
skip = "uv.lock,*.ipynb,*.lock,data/*"
ignore-words-list = "te,nd,fo"
